chore: move register_debug_type to new module to simplify assembler.rs
